Refactoring Login form
Rename Login component for clarity
Remove duplicated header
