+++ src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { UserprofileComponent } from './layout/userprofile/userprofile.component
 import { AccountComponent } from './login/account/account.component';
 import { ValidateemailComponent } from './login/account/validateemail/validateemail.component';
 import { LoginComponent } from './login/login.component';
-import { MainPageComponent } from './login/main-page/main-page.component';
+import { LoginFormComponent } from './login/login-form/login-form.component';
 import { PasswordConfirmationComponent } from './login/password-reset/password-confirmation/password-confirmation.component';
 import { PasswordResetComponent } from './login/password-reset/password-reset.component';
 import PathConstants from '../environments/paths';
@@ -17,7 +17,7 @@ const routes: Routes = [
     children: [
       {
         path: '',
-        component: MainPageComponent,
+        component: LoginFormComponent,
         pathMatch: 'full',
       },
       {
