+++ src/app/app.module.ts
@@ -5,37 +5,36 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { LoginComponent } from './login/login.component';
-import { HeaderComponent } from './login/header/header.component';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import { MainPageComponent } from './login/main-page/main-page.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { LoginFormComponent } from './login/login-form/login-form.component';
 import { PasswordResetComponent } from './login/password-reset/password-reset.component';
 import { PasswordConfirmationComponent } from './login/password-reset/password-confirmation/password-confirmation.component';
 import { AccountComponent } from './login/account/account.component';
 import { ValidateemailComponent } from './login/account/validateemail/validateemail.component';
 import { LayoutComponent } from './layout/layout.component';
-import { MainheaderComponent } from './layout/mainheader/mainheader.component';
+import { HeaderComponent } from './layout/header/header.component';
 import { RouterModule } from '@angular/router';
 import { SidebarComponent } from './layout/sidebar/sidebar.component';
 import { UserprofileComponent } from './layout/userprofile/userprofile.component';
-import {AuthInterceptor} from "./services/auth/auth.interceptor";
-import {HTTP_INTERCEPTORS, HttpClientModule} from "@angular/common/http";
+import { AuthInterceptor } from './services/auth/auth.interceptor';
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { AlertComponent } from './shared/alert/alert.component';
 import { LoadingSpinnerComponent } from './shared/loading-spinner/loading-spinner.component';
-import {environment} from "../environments/environment";
-import {LoggerModule} from "ngx-logger";
+import { environment } from '../environments/environment';
+import { LoggerModule } from 'ngx-logger';
 
 @NgModule({
   declarations: [
     AppComponent,
     LoginComponent,
-    HeaderComponent,
-    MainPageComponent,
+
+    LoginFormComponent,
     PasswordResetComponent,
     PasswordConfirmationComponent,
     AccountComponent,
     ValidateemailComponent,
     LayoutComponent,
-    MainheaderComponent,
+    HeaderComponent,
     SidebarComponent,
     UserprofileComponent,
     AlertComponent,
@@ -49,9 +48,11 @@ import {LoggerModule} from "ngx-logger";
     ReactiveFormsModule,
     FormsModule,
     HttpClientModule,
-    LoggerModule.forRoot({level: environment.logLevel})
+    LoggerModule.forRoot({ level: environment.logLevel }),
   ],
-  providers: [{ provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true}],
-  bootstrap: [AppComponent]
+  providers: [
+    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+  ],
+  bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
