+++ src/app/layout/header/header.component.ts
@@ -1,13 +1,15 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { AuthService } from '../../services/auth/auth.service';
 
 @Component({
   selector: 'app-mainheader',
-  templateUrl: './mainheader.component.html',
-  styleUrls: ['./mainheader.component.css'],
+  templateUrl: './header.component.html',
+  styleUrls: ['./header.component.css'],
 })
-export class MainheaderComponent implements OnInit {
+export class HeaderComponent implements OnInit {
+  @Input() hideSignInButton = false;
+
   buttonMsg = 'Sign in';
 
   constructor(private router: Router, private authService: AuthService) {}
