--- src/app/layout/mainheader/mainheader.component.spec.ts
@@ -1,46 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { MainheaderComponent } from './mainheader.component';
-import { RouterTestingModule } from '@angular/router/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import {
-  LoggerConfig,
-  NGXLogger,
-  NGXLoggerHttpService,
-  NgxLoggerLevel,
-  NGXMapperService,
-} from 'ngx-logger';
-import {
-  NGXLoggerHttpServiceMock,
-  NGXMapperServiceMock,
-} from 'ngx-logger/testing';
-import { DatePipe } from '@angular/common';
-
-describe('MainheaderComponent', () => {
-  let component: MainheaderComponent;
-  let fixture: ComponentFixture<MainheaderComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [MainheaderComponent],
-      imports: [RouterTestingModule, HttpClientTestingModule],
-      providers: [
-        NGXLogger,
-        { provide: NGXLoggerHttpService, useClass: NGXLoggerHttpServiceMock },
-        { provide: NGXMapperService, useClass: NGXMapperServiceMock },
-        { provide: LoggerConfig, useValue: { level: NgxLoggerLevel.ERROR } },
-        DatePipe,
-      ],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(MainheaderComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
