+++ src/app/login/account/account.component.html
@@ -1,4 +1,4 @@
-<app-header></app-header>
+<app-mainheader></app-mainheader>
 <div class="container-fluid">
     <form>
         <div class="row movingGradient no-gutters">
