--- src/app/login/header/header.component.html
@@ -1,5 +1,0 @@
-<nav class="navbar" aria-label="Navigation bar">
-    <div class="container-fluid">
-      <a class="navbar-brand mb-0 h1 leftHeader movingGradient" href="" aria-label="Return back to Utopia home page">UTOPIA</a>
-    </div>
-</nav>
