+++ src/app/login/login-form/login-form.component.html
@@ -1,4 +1,4 @@
-<app-header></app-header>
+<app-mainheader [hideSignInButton]="true"></app-mainheader>
 <div class="container-fluid componentContainer">
   <app-alert
     *ngIf="errorMsg"
