+++ src/app/login/password-reset/password-reset.component.html
@@ -1,4 +1,4 @@
-<app-header></app-header>
+<app-mainheader [hideSignInButton]="true"></app-mainheader>
 <div class="container-fluid componentContainer">
     <div class="row">
         <div class="col d-flex justify-content-center firstRow">
